requirements.txt
asgiref==3.7.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0
django-filter==2.2.0
django-import-export==2.0.2
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.0.1
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
packaging==23.1
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
tablib==1.1.0
typing_extensions==4.7.1
uritemplate==4.1.1
xlrd==1.2.0
xlwt==1.3.0
